Added PCI, generalized table function
I have started work on PCI. The enumerate function is able to
successfully, albeit basically, read the configuration space.
I still need to work out access of the configuration space from
the IO ports. The enumeration function needs to become more complex, for
now it just lists information it finds about the 256 devices in the
first bus segment on the first bus of the segment. I made a small to
change to ACPI, this is so that the function to get the data of a table
is more generalized, a caller now passes a table ID and a return pointer
instead of just a return pointer and calling a specific function to get
the information of that table.