Begin work on switching to a process
A program is now successfully maped into memory, has the kernel mapped
in for it as well, and is given a little bit of memory to work with. I
have begun work on writing the process_switch function. It does manage
to switch out the page tables, thouhg it does not manage to switch to
the process. This is because if you switch to the page tables you can no
longer access all of the nice objects made by the allocator. If you
switch first then you are hoping the timer interrupt happens just after
you change the page tables. Therefore, the only real way I see forward
currently is to make it so that the page tables are also flipped on the
timer interrupt (if signaled to do so) and somehow have a hook there to
write out the registers to an actual thread instead just writing to a
processor descriptor's context.