Simplify thread getting, current_process, RFLAGs
I have simplified the thread getting code in the process. It now no
longer needs the current thread explicitly given to it. It will assume
it from the processor's descriptor structure. There has also been
additional condition added to return NULL - if the current thread is the
same as the one that is to be returned. I have also corrected the
initial value of RFLAGs to set the Interrupt Enable bit.