Future proofing
The processor descriptor has received a pointer to scheduler metadata. This will allow schedulers to be per processor. The thread structure now also contains a struct to keep track of profiling data and a uint32_t to sum up the profiling data. I hope that this will be useful in the future for optimization in the schedulers; this comes from a little bit of reading about NUMA. The declaration for the holding process has been moved into global.h.