Comply with Kmm, pager_unmap return type
I have modified the definition for the pager_unmap function to return a uinptr_t (the first physical address that the virtual address was mapped to). This is so that contiguous allocations can be freed. This may be further changed to return a linked list of all the addresses mapped in that region in the form {base, size}.