Standardize floats and context
I have made a standard interface for floats and implemented a new context header file which just includes the correct context header from the specific architecture. In SMP, the state of the processor has been changed from just being its registers to being the new context structure, as this structure allows for storing of additional information which does not need to be accounted for in assembly.