Remove context writing / saving
These functions are obsolete as I do not foresee the need to specifically say that this processor should write out its register state or read in some register state and so on. This allows for removal of the context and the reservation of the first two bits of the flags structure. The locks have also been removed in to hopefully do timer setting atomically.