ACPI: generalize
A set of functions have been declared for an ACPI implementation to
define under include/arch/acpi. The implementation specific parts have
been moved to acpi/XXX/ (in this case acpi/uacpi). This is not to be
confused with the uacpi folder that is located in Arctan-OS/Kernel, that
is the actual implementation of uACPI.