pager: remove silly source bool
The source parameter should have been a bool instead of int, but that is
not the biggest problem. The biggest problem is that it is in general
confusing. I have chosen to remove it and instead allow the caller to
specify the destination and source tables. This should offer far more
versatility.