context: declare helpful functions
These functions are to abstract the context saving and loading.
Further, there is another function that is to be used for setting
processor features. This is because I intend to migrate the KernelMeta
structure / processor feature detection to be per processor. This will
leave a pathway open in the future for NUMA.