pci: fix logical issues
I made a bunch of logical errors when originally writing this. I have
since fixed those. There still may be more issues that need resolving,
but now it isn't so blatantly wrong. This has been done by making a
header meta structure containing a pointer to the header itself. This
allows for both functions to use the same structure.