pci: add enum, remove functions, pci_free_header
I added an enum to make it easier to switch between the various types of
PCI header. Further, I removed the pci_get_status and pci_set_command
functions as I don't see them as being needed. I added a function also
to clean up a header meta structure. The enumerate function was also
slightly changed such that it uses the updated signature for the
init_pci_resource function and uses the new functions to create the PCI
meta header. I will be getting to work on accomodating PCI-PCI bridges.